+++ ext/parquet/build.rs
@@ -1,5 +1,5 @@
 pub fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let _rb_env = rb_sys_env::activate()?;
+    rb_sys_env::activate()?;
 
     Ok(())
 }